Remove compiler warnings Fixed wrong table flags type in ndbcluster that caused previous commit to fail

+++ sql/mysql_priv.h
@@ -1518,7 +1518,7 @@ extern bool opt_using_transactions, mysqld_embedded;
 extern bool using_update_log, opt_large_files, server_id_supplied;
 extern bool opt_update_log, opt_bin_log, opt_error_log;
 extern my_bool opt_log, opt_slow_log;
-extern uint log_output_options;
+extern ulong log_output_options;
 extern my_bool opt_log_queries_not_using_indexes;
 extern bool opt_disable_networking, opt_skip_show_db;
 extern my_bool opt_character_set_client_handshake;
